Get real literals for nations
