+++ src/pykeen/datasets/generate_literals.py
@@ -6,7 +6,6 @@ import click
 import torch
 
 import pykeen.nn
-from pykeen.datasets.tnl import LITERALS_PATH
 from pykeen.models.base import EntityRelationEmbeddingModel
 from pykeen.pipeline import PipelineResult, pipeline
 
@@ -14,13 +13,21 @@ from pykeen.pipeline import PipelineResult, pipeline
 @click.command()
 @click.option('--features', type=int, default=2, show_default=True)
 @click.option('--seed', type=int, default=2, show_default=True)
-def generate(features: int, seed: int):
+@click.option('--output', type=click.File('w'))
+@click.option('--dataset', required=True)
+def generate(features: int, seed: int, output):
     """Generate random literals for the Nations dataset."""
+    for h, r, t in generate_literals(features=features, seed=seed):
+        print(h, r, t, sep='\t', file=output)
+
+
+def generate_literals(dataset: str, features: int, seed: int, model: str = 'rotate'):
+    """Generate literals or the given dataset."""
     generator = torch.Generator(seed)
 
     pipeline_result: PipelineResult = pipeline(
-        dataset='nations',
-        model='rotate',
+        dataset=dataset,
+        model=model,
         training_kwargs=dict(
             num_epochs=120,
         ),
@@ -41,7 +48,4 @@ def generate(features: int, seed: int):
         feature += torch.normal(mean=0, std=1, size=feature.size(), generator=generator)
         for (_, label), value in zip(sorted(tf.entity_id_to_label.items()), feature):
             rows.append((label, f'feature{i}', value.item()))
-
-    with open(LITERALS_PATH, 'w') as file:
-        for h, r, t in rows:
-            print(h, r, t, sep='\t', file=file)
+    return rows
