+++ src/pykeen/datasets/nations/literals.txt
@@ -1,0 +1,26 @@
+brazil	area	8514877
+burma	area	676578
+china	area	9596960
+cuba	area	110860
+egypt	area	1001450
+india	area	3287263
+indonesia	area	1904569
+israel	area	20770
+jordan	area	89342
+netherlands	area	41543
+poland	area	312685
+uk	area	243610
+usa	area	9826675
+brazil	population	204259812
+burma	population	56320206
+china	population	1367485388
+cuba	population	11031433
+egypt	population	88487396
+india	population	1251695584
+indonesia	population	255993674
+israel	population	8049314
+jordan	population	8117564
+netherlands	population	16947904
+poland	population	38562189
+uk	population	64088222
+usa	population	321368864
