--- src/pykeen/datasets/tnl/literals.txt
@@ -1,28 +1,0 @@
-brazil	feature0	-0.23296403884887695
-burma	feature0	-3.188920021057129
-china	feature0	-0.9473296999931335
-cuba	feature0	0.5893441438674927
-egypt	feature0	-0.7200393080711365
-india	feature0	-2.8713979721069336
-indonesia	feature0	-2.391174793243408
-israel	feature0	-4.305821418762207
-jordan	feature0	-0.631758451461792
-netherlands	feature0	-1.1200616359710693
-poland	feature0	-3.405829668045044
-uk	feature0	0.6276875734329224
-usa	feature0	-0.3950362801551819
-ussr	feature0	4.355489730834961
-brazil	feature1	1.2958908081054688
-burma	feature1	-2.1253457069396973
-china	feature1	-0.7485345005989075
-cuba	feature1	-1.638179063796997
-egypt	feature1	1.156923770904541
-india	feature1	-1.6402899026870728
-indonesia	feature1	-1.4675965309143066
-israel	feature1	1.9007127285003662
-jordan	feature1	-1.3840017318725586
-netherlands	feature1	-1.5874662399291992
-poland	feature1	-3.2188329696655273
-uk	feature1	2.2382407188415527
-usa	feature1	-0.2731866240501404
-ussr	feature1	3.63517427444458
